versions/v0.05/requirements.txt
# GUI Framework
kivy>=2.2.1
pillow>=10.0.0  # Required for Kivy image handling

# Web Scraping & Networking
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.3  # Faster parser for BeautifulSoup

# Database
sqlite3  # Usually included with Python

# Logging & Console
colorama>=0.4.6
rich>=13.3.5  # Enhanced console output

# Development Tools
python-dotenv>=1.0.0
black>=23.7.0
pylint>=2.17.5
mypy>=1.5.1  # Type checking

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-kivy>=0.3.0  # Kivy-specific testing support
pytest-asyncio>=0.21.1  # For async tests

# Documentation
sphinx>=7.1.2
sphinx-rtd-theme>=1.3.0

# Performance Optimizations
cython>=3.0.0  # For better Kivy performance
pillow-simd>=9.0.0  # Faster image processing (optional)

# Cache Management
diskcache>=5.4.0  # Efficient disk caching

# Type Hints
typing-extensions>=4.7.1  # Enhanced type hinting support

# Development
watchdog>=3.0.0  # For development auto-reload

# Android Support
buildozer
android